api/requirements.txt
fastapi
passlib[bcrypt]
email-validator
pyjwt
uvicorn
SQLAlchemy
sqlalchemy_utils
psycopg2